refactory: criando o componente de wrapper e ajuste no botoes
